Refactoring common back into rest
