+++ gaffer-as-a-service/gaas-rest/src/main/java/uk/gov/gchq/gaffer/gaas/model/GaaSGraph.java
@@ -16,7 +16,6 @@
 
 package uk.gov.gchq.gaffer.gaas.model;
 
-import uk.gov.gchq.gaffer.common.model.v1.RestApiStatus;
 import java.util.List;
 
 public class GaaSGraph {
