+++ gaffer-as-a-service/gaas-rest/src/main/java/uk/gov/gchq/gaffer/gaas/model/Gaffer.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package uk.gov.gchq.gaffer.common.model.v1;
+package uk.gov.gchq.gaffer.gaas.model;
 
 import com.google.gson.annotations.SerializedName;
 import io.kubernetes.client.common.KubernetesObject;
@@ -22,7 +22,8 @@ import io.kubernetes.client.openapi.models.V1ObjectMeta;
 import org.apache.commons.lang3.builder.EqualsBuilder;
 import org.apache.commons.lang3.builder.HashCodeBuilder;
 import org.apache.commons.lang3.builder.ToStringBuilder;
-import uk.gov.gchq.gaffer.common.util.Constants;
+import uk.gov.gchq.gaffer.gaas.util.Constants;
+
 
 /**
  * Value object for Gaffer
