+++ gaffer-as-a-service/gaas-rest/src/main/java/uk/gov/gchq/gaffer/gaas/model/GafferSpec.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package uk.gov.gchq.gaffer.common.model.v1;
+package uk.gov.gchq.gaffer.gaas.model;
 
 
 import com.google.common.collect.Lists;
