+++ gaffer-as-a-service/gaas-rest/src/main/java/uk/gov/gchq/gaffer/gaas/services/GetGaaSGraphConfigsService.java
@@ -18,11 +18,12 @@ package uk.gov.gchq.gaffer.gaas.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import uk.gov.gchq.gaffer.common.model.v1.GafferSpec;
 import uk.gov.gchq.gaffer.federatedstore.FederatedStore;
 import uk.gov.gchq.gaffer.gaas.exception.GaaSRestApiException;
 import uk.gov.gchq.gaffer.gaas.model.GafferConfigSpec;
+import uk.gov.gchq.gaffer.gaas.model.GafferSpec;
 import uk.gov.gchq.gaffer.gaas.util.GafferSpecConfigsLoader;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
