+++ gaffer-as-a-service/gaas-rest/src/test/java/uk/gov/gchq/gaffer/gaas/handlers/DeploymentHandlerTest.java
@@ -38,29 +38,20 @@ import org.slf4j.LoggerFactory;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.core.env.Environment;
 import org.springframework.test.util.ReflectionTestUtils;
-import uk.gov.gchq.gaffer.common.model.v1.Gaffer;
-import uk.gov.gchq.gaffer.common.model.v1.GafferSpec;
 import uk.gov.gchq.gaffer.gaas.factories.IKubernetesObjectFactory;
+import uk.gov.gchq.gaffer.gaas.model.Gaffer;
+import uk.gov.gchq.gaffer.gaas.model.GafferSpec;
 import uk.gov.gchq.gaffer.gaas.util.UnitTest;
+
 import java.util.HashMap;
 import java.util.List;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotEquals;
-import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.junit.jupiter.api.Assertions.assertTrue;
+
+import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
-import static uk.gov.gchq.gaffer.common.util.Constants.WORKER_HELM_REPO;
-import static uk.gov.gchq.gaffer.common.util.Constants.WORKER_IMAGE;
-import static uk.gov.gchq.gaffer.common.util.Constants.WORKER_IMAGE_PULL_POLICY;
-import static uk.gov.gchq.gaffer.common.util.Constants.WORKER_NAMESPACE;
-import static uk.gov.gchq.gaffer.common.util.Constants.WORKER_RESTART_POLICY;
-import static uk.gov.gchq.gaffer.common.util.Constants.WORKER_SERVICE_ACCOUNT_NAME;
-import static uk.gov.gchq.gaffer.gaas.util.Constants.CONFIG_NAME_KEY;
-import static uk.gov.gchq.gaffer.gaas.util.Constants.DESCRIPTION_KEY;
-import static uk.gov.gchq.gaffer.gaas.util.Constants.GRAPH_ID_KEY;
+import static uk.gov.gchq.gaffer.gaas.util.Constants.*;
 
 @EnableKubernetesMockClient(crud = true)
 @UnitTest
