+++ gaffer-as-a-service/gaas-rest/src/test/java/uk/gov/gchq/gaffer/gaas/model/GafferSpecTest.java
@@ -15,7 +15,7 @@
  */
 
 
-package uk.gov.gchq.gaffer.common.model.v1;
+package uk.gov.gchq.gaffer.gaas.model;
 
 import org.junit.jupiter.api.Test;
