+++ gaffer-as-a-service/gaas-rest/src/test/java/uk/gov/gchq/gaffer/gaas/util/GafferSpecConfigsLoaderTest.java
@@ -19,7 +19,7 @@ package uk.gov.gchq.gaffer.gaas.util;
 import com.google.gson.Gson;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import uk.gov.gchq.gaffer.common.model.v1.GafferSpec;
+import uk.gov.gchq.gaffer.gaas.model.GafferSpec;
 import uk.gov.gchq.gaffer.gaas.exception.GaaSRestApiException;
 
 import java.util.HashMap;
